Add control for mate happiness distance, other cosmetic changes

+++ src/bankViewer/Insert.hh
@@ -24,7 +24,7 @@ public:
   int m_length;
   int m_actual;
 
-  static int MAXSTDEV;
+  static float MAXSTDEV;
 
   int m_arc;
   int m_brc;
